Changed models
Changed models

+++ api/Models/Movement.cs
@@ -2,9 +2,9 @@
 {
     public class Movement : Common
     {
-        public required float XValue { get; set; }
-        public required float YValue { get; set; }
-        public required float ZValue { get; set; }
+        public required int XValue { get; set; }
+        public required int YValue { get; set; }
+        public required int ZValue { get; set; }
         public string sitSmartDeviceId { get; set; }
         public SitSmartDevice sitSmartDevice { get; set; }
     }
